requirements.txt
beautifulsoup4==4.11.1
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.0
streamlit==1.10.0
